requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
black==23.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
exceptiongroup==1.1.1
fire==0.5.0
flake8==6.0.0
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.32
idna==3.4
iniconfig==2.0.0
isort==5.12.0
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
openai==0.27.6
packaging==23.1
pathspec==0.11.1
platformdirs==3.5.0
pluggy==1.0.0
prompt-toolkit==3.0.39
pycodestyle==2.10.0
pyflakes==3.0.1
Pygments==2.15.1
pytest==7.3.1
pytest-mock==3.10.0
pytest-repeat==0.9.1
pytest-reportlog==0.4.0
python-dotenv==1.0.0
regex==2023.6.3
requests==2.29.0
ruff==0.0.270
six==1.16.0
smmap==5.0.0
termcolor==2.3.0
tiktoken==0.4.0
tomli==2.0.1
tqdm==4.65.0
urllib3==1.26.15
wcwidth==0.2.6
yarl==1.9.2
